+++ backend/src/main/java/com/ProTeen/backend/community/dto/ResponseDTO.java
@@ -1,0 +1,18 @@
+package com.ProTeen.backend.community.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+import java.util.List;
+
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+@Data
+public class ResponseDTO<T> {
+    private String error;
+    private PageDTO page;
+    private List<T> data;
+}
