+++ backend/src/main/java/com/ProTeen/backend/community/model/ImageEntity.java
@@ -1,0 +1,27 @@
+package com.ProTeen.backend.community.model;
+
+import jakarta.persistence.*;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Builder
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+@Entity
+@Table(name = "image")
+public class ImageEntity {
+    @Id
+    @Column(name = "id")
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @Column(name = "imgPath")
+    private String imgPath;
+
+    @ManyToOne
+    @JoinColumn(name = "board_id")
+    private BoardEntity board;
+}
