+++ backend/src/main/java/com/ProTeen/backend/community/repository/CommentRepository.java
@@ -1,0 +1,11 @@
+package com.ProTeen.backend.community.repository;
+
+import com.ProTeen.backend.community.model.CommentEntity;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import java.util.List;
+@Repository
+public interface CommentRepository extends JpaRepository<CommentEntity, Long> {
+    List<CommentEntity> findByAuthor(String author);
+}
