+++ backend/src/main/java/com/ProTeen/backend/community/repository/ImageRepository.java
@@ -1,0 +1,9 @@
+package com.ProTeen.backend.community.repository;
+
+import com.ProTeen.backend.community.model.ImageEntity;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface ImageRepository extends JpaRepository<ImageEntity, Long> {
+}
