--- backend/src/main/java/com/ProTeen/backend/config/AppConfig.java
@@ -1,10 +1,0 @@
-package com.ProTeen.backend.config;
-
-import com.ProTeen.backend.service.BoardService;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-public class AppConfig {
-
-}
