--- backend/src/main/java/com/ProTeen/backend/dto/PageDTO.java
@@ -1,22 +1,0 @@
-package com.ProTeen.backend.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import java.util.List;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-public class PageDTO {
-
-    private List<BoardDTO.Summary> content;
-    private int pageNo;
-    private int pageSize;
-    private long totalElements;
-    private int totalPages;
-
-}
