--- backend/src/main/java/com/ProTeen/backend/dto/ResponseDTO.java
@@ -1,17 +1,0 @@
-package com.ProTeen.backend.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import java.util.List;
-
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-@Data
-public class ResponseDTO<T> {
-    private String error;
-    private List<T> data;
-}
