--- backend/src/main/java/com/ProTeen/backend/model/ImageEntity.java
@@ -1,30 +1,0 @@
-package com.ProTeen.backend.model;
-
-import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-import org.hibernate.annotations.GenericGenerator;
-
-import java.time.LocalDateTime;
-
-@Builder
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-@Entity
-@Table(name = "image")
-public class ImageEntity {
-    @Id
-    @Column(name = "id")
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(name = "imgPath")
-    private String imgPath;
-
-    @ManyToOne
-    @JoinColumn(name = "board_id")
-    private BoardEntity board;
-}
