--- backend/src/main/java/com/ProTeen/backend/model/UserEntity.java
@@ -1,50 +1,0 @@
-package com.ProTeen.backend.model;
-
-import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-import org.hibernate.annotations.GenericGenerator;
-import org.springframework.format.annotation.DateTimeFormat;
-
-import java.util.ArrayList;
-import java.util.List;
-
-@Data
-@Entity
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-@Table(name = "USERS", uniqueConstraints = {@UniqueConstraint(columnNames = {"userId", "nickName"})})
-public class UserEntity {
-
-    @Id
-    @GeneratedValue(generator = "system-uuid")
-    @GenericGenerator(name = "system-uuid", strategy = "uuid")
-    private String id;
-
-    @Column(nullable = false)
-    private String userId;
-
-    @Column(nullable = false)
-    private String userPassword;
-
-    @Column(nullable = false)
-    private String nickname;
-
-    @Column(nullable = false)
-    private String createdDate;
-
-    @Column
-    private String role;
-
-//    @OneToMany(mappedBy = "user")
-//    private List<Article> articles = new ArrayList<>();
-//
-//    @OneToMany(mappedBy = "user")
-//    private List<Comment> comments = new ArrayList<>();
-//
-//    @OneToMany(mappedBy = "user")
-//    private List<Test> test = new ArrayList<>();
-}
