--- backend/src/main/java/com/ProTeen/backend/repository/CommentRepository.java
@@ -1,11 +1,0 @@
-package com.ProTeen.backend.repository;
-
-import com.ProTeen.backend.model.CommentEntity;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.List;
-@Repository
-public interface CommentRepository extends JpaRepository<CommentEntity, Long> {
-    List<CommentEntity> findByAuthor(String author);
-}
