--- backend/src/main/java/com/ProTeen/backend/repository/ImageRepository.java
@@ -1,9 +1,0 @@
-package com.ProTeen.backend.repository;
-
-import com.ProTeen.backend.model.ImageEntity;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface ImageRepository extends JpaRepository<ImageEntity, Long> {
-}
