--- backend/src/main/java/com/ProTeen/backend/repository/UserRepository.java
@@ -1,13 +1,0 @@
-package com.ProTeen.backend.repository;
-
-
-import com.ProTeen.backend.model.UserEntity;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-
-@Repository
-public interface UserRepository extends JpaRepository<UserEntity, String> {
-    UserEntity findByUserId(String userId);
-    Boolean existsByUserId(String userId);
-}
