+++ backend/src/main/java/com/ProTeen/backend/user/config/AppConfig.java
@@ -1,0 +1,8 @@
+package com.ProTeen.backend.user.config;
+
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+public class AppConfig {
+
+}
