+++ backend/src/main/java/com/ProTeen/backend/user/config/SecurityConfig.java
@@ -1,0 +1,58 @@
+package com.ProTeen.backend.user.config;
+
+import com.ProTeen.backend.user.security.JwtAuthenticationFilter;
+import lombok.AllArgsConstructor;
+import org.springframework.http.HttpMethod;
+import org.springframework.web.filter.CorsFilter;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.security.config.annotation.web.builders.HttpSecurity;
+import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
+import org.springframework.security.config.http.SessionCreationPolicy;
+import org.springframework.security.web.SecurityFilterChain;
+import org.springframework.security.web.header.writers.frameoptions.XFrameOptionsHeaderWriter;
+import org.springframework.security.web.util.matcher.AntPathRequestMatcher;
+
+@Configuration
+@EnableWebSecurity
+@AllArgsConstructor
+public class SecurityConfig {
+    @Autowired
+    private JwtAuthenticationFilter jwtAuthenticationFilter;
+
+
+    @Bean
+    public SecurityFilterChain filterChain(HttpSecurity http) throws Exception {
+
+        http.cors()
+                .and()
+                .httpBasic().disable() // basic 인증 disable
+                // 세션 기반이 아님을 선언
+                .sessionManagement().sessionCreationPolicy(SessionCreationPolicy.STATELESS)
+                .and()
+                .authorizeHttpRequests()
+                .requestMatchers("/main/**").permitAll()
+                .requestMatchers(HttpMethod.GET).permitAll()
+                .requestMatchers(AntPathRequestMatcher.antMatcher("/h2-console/**")).permitAll()
+                .and()
+                .authorizeHttpRequests()
+                .anyRequest()
+                .authenticated() // 이외의 경로는 인증
+                .and()
+                .csrf()
+                .disable().headers().addHeaderWriter(new XFrameOptionsHeaderWriter(
+                        XFrameOptionsHeaderWriter.XFrameOptionsMode.SAMEORIGIN));
+
+        // filter 등록
+        // 매 요청마다
+        // CorsFilter 실행한 후에
+        // jwtAuthenticationFilter 실행
+        http.addFilterAfter(
+                jwtAuthenticationFilter,
+                CorsFilter.class
+        );
+
+        return http.build();
+    }
+}
