+++ backend/src/main/java/com/ProTeen/backend/user/dto/UserDTO.java
@@ -1,0 +1,18 @@
+package com.ProTeen.backend.user.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+public class UserDTO {
+    private String token;
+    private String userId;
+    private String nickname;
+    private String password;
+    private String id;
+}
