+++ backend/src/main/java/com/ProTeen/backend/user/handler/LoginSuccessHandler.java
@@ -1,0 +1,24 @@
+package com.ProTeen.backend.user.handler;
+
+import jakarta.servlet.ServletException;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+import jakarta.servlet.http.HttpSession;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.web.authentication.AuthenticationSuccessHandler;
+
+import java.io.IOException;
+
+@Slf4j
+public class LoginSuccessHandler implements AuthenticationSuccessHandler {
+    @Override
+    public void onAuthenticationSuccess(HttpServletRequest request, HttpServletResponse response,
+                                        Authentication authentication)
+            throws IOException, ServletException {
+        log.info("handler invoked");
+        HttpSession session = request.getSession();
+        session.setAttribute("greeting", authentication.getName() + "님 반가워요");
+        response.sendRedirect("/main");
+    }
+}
